+++ frontend/svelte/src/lib/components/ui/Wizard/WizardPanel.svelte
